refactor function name and files
