+++ webapp/forbes/patient_report.php
@@ -1,5 +1,6 @@
 <?php
 // The user is redirected here from login.php.
+include 'data_retriever.php';
 function redirect_user($page)
 {
     // Start defining the URL...
@@ -9,95 +10,6 @@ function redirect_user($page)
     // Redirect the user:
     header("Location: $url");
     exit(); // Quit the script.
-}
-function get_cost($type, $rid)
-{
-    require "../mysqli_connect.php"; // Connect to the db.
-
-    $pid = $_GET["id"];
-    // get days in day2-x
-    $qDay = "SELECT DATEDIFF(IFNULL(DATE_SUB(checkout, INTERVAL 1 DAY), CURDATE()), checkin) as 'day' FROM patients WHERE patient_id='$pid'";
-    $rDay = @mysqli_query($dbc, $qDay);  // run query
-    if (mysqli_num_rows($rDay) == 1) { // ok
-        $rowDay = mysqli_fetch_assoc($rDay);
-        $day = $rowDay['day'];
-        if($day > 1) {
-            $day -= 1;
-        } else {
-            $day = 0;
-        }
-    } else {
-        return "error";
-    }
-
-    // get time spent 
-    if ($type=='sg') {
-        $qTime = "SELECT SUM(freq * time_duration) as 'total' FROM reports WHERE (patient_id='$pid' AND role_id=$rid AND activity_id<=6)";
-    } else if ($type=='po') {
-        $qTime = "SELECT
-                    (SELECT SUM(freq * time_duration) as 't' FROM reports WHERE patient_id='$pid' AND role_id=$rid AND activity_id>6 AND activity_day!='d') +
-                    (SELECT $day*SUM(freq * time_duration) as 't' FROM reports WHERE patient_id='$pid' AND role_id=$rid AND activity_id>6 AND activity_day='d')
-                  AS 'total';";     
-    } else {
-        $qTime = "SELECT
-                    (SELECT SUM(freq * time_duration) as 't' FROM reports WHERE patient_id='$pid' AND role_id=$rid AND activity_day!='d') +
-                    (SELECT IFNULL($day*SUM(freq * time_duration),0) as 't' FROM reports WHERE patient_id='$pid' AND role_id=$rid AND activity_day='d')
-                  AS 'total';"; 
-    }  
-    $rTime = @mysqli_query($dbc, $qTime);  // run query
-    // get salary
-    $qPay = "SELECT salary FROM roles WHERE role_id=$rid";   
-    $rPay = @mysqli_query($dbc, $qPay);  // run query 
-    if (mysqli_num_rows($rTime) == 1 & mysqli_num_rows($rPay) == 1) { // ok
-        $rowTime = mysqli_fetch_assoc($rTime);
-        $rowPay = mysqli_fetch_assoc($rPay);
-        $result = $rowTime['total'] * $rowPay['salary'] / 124800; // 52 week * 40 hours * 60 minuts
-        return number_format($result,0);
-    } else {
-        return "error";
-    }
-    mysqli_close($dbc);
-}
-
-
-function get_value($graph, $input) {
-    require ('../mysqli_connect.php'); // Connect to the db.
-    
-    $pid = $_GET["id"];
-    if ($graph == "cost") {
-        if ($input == "hospital") { // get hospital wide cost
-            $q = "SELECT AVG(d.total) AS 'result' FROM
-                    (SELECT a.patient_id, SUM(a.freq * a.time_duration * b.salary / 124800 *
-		                    (CASE
-			                     WHEN a.activity_day = 'd' THEN c.d
-			                     ELSE 1
-                             END)) AS 'total' 
-                     FROM   reports a
-		             INNER JOIN (SELECT role_id, salary FROM roles ) b ON a.role_id=b.role_id
-                     INNER JOIN (SELECT patient_id, DATEDIFF(checkout, checkin)-2 as 'd' FROM patients WHERE checkout IS NOT NULL) c ON a.patient_id=c.patient_id
-                     GROUP BY patient_id) d";
-        } else { // get reimbursement
-            
-        }
-    } else {
-        if ($input == "patient") { // get patient stay
-            $q = "SELECT DATEDIFF(IFNULL(checkout, CURDATE()), checkin) as 'result' FROM patients WHERE patient_id='$pid'";
-        } else if ($input == "hospital") { // get hospital wide stay
-            $q = "SELECT AVG(DATEDIFF(checkout, checkin)) as 'result' FROM patients WHERE checkout IS NOT NULL";
-        } else { // get national
-            
-        }
-    }
-    
-    $r = @mysqli_query($dbc, $q);  // run query
-    if (mysqli_num_rows($r) == 1) { // ok
-        $row = mysqli_fetch_array($r);
-        $result = $row['result'];
-        return $result;
-    } else {
-        return "error:";
-    }
-    mysqli_close($dbc);
 }
 
 if (!isset($_COOKIE['email'])) { // If no cookie is present, redirect:
@@ -194,23 +106,23 @@ if (!isset($_COOKIE['email'])) { // If no cookie is present, redirect:
                                         <tbody>
                                             <tr>
                                                 <td>Anesthesiologist</td>
-                                                <td><span class="sg" name="sg-an"><?php echo get_cost('to',1) ?></span></td>
+                                                <td><span class="sg" name="sg-an"><?php echo get_patient_cost('to',1) ?></span></td>
                                             </tr>
                                             <tr>
                                                 <td>Cardiovascular Surgeon</td>
-                                                <td><span class="sg" name="sg-cs"><?php echo get_cost('to',3) ?></span></td>
+                                                <td><span class="sg" name="sg-cs"><?php echo get_patient_cost('to',3) ?></span></td>
                                             </tr> 
                                             <tr>
                                                 <td>Physician Assistant</td>
-                                                <td><span class="sg" name="sg-pa"><?php echo get_cost('sg',10) ?></span></td>
+                                                <td><span class="sg" name="sg-pa"><?php echo get_patient_cost('sg',10) ?></span></td>
                                             </tr>                                            
                                             <tr>
                                                 <td>Registered Nurse</td>
-                                                <td><span class="sg" name="sg-rn"><?php echo get_cost('sg',8) ?></span></td>                                                     
+                                                <td><span class="sg" name="sg-rn"><?php echo get_patient_cost('sg',8) ?></span></td>
                                             </tr>
                                             <tr>
                                                 <td>Scrub Tech</td>
-                                                <td><span class="sg" name="sg-st"><?php echo get_cost('sg',13) ?></span></td>
+                                                <td><span class="sg" name="sg-st"><?php echo get_patient_cost('sg',13) ?></span></td>
                                             </tr>  
                                             <tr>
                                                 <td><strong>Total:</strong></td>
@@ -229,7 +141,7 @@ if (!isset($_COOKIE['email'])) { // If no cookie is present, redirect:
                                         <tbody>
                                             <tr>
                                             <td>All Direct Material</td>
-                                            <td><span name="sg-dm">1</span></td>                                                     
+                                            <td><span name="sg-dm"><?php echo get_dm() ?></span></td>
                                             </tr>
                                         </tbody>
                                     </table>
@@ -244,7 +156,7 @@ if (!isset($_COOKIE['email'])) { // If no cookie is present, redirect:
                                         <tbody>
                                             <tr>
                                             <td>All Overhead</td>
-                                            <td><span name="sg-oh">1000</span></td>                                                     
+                                            <td><span name="sg-oh"><?php echo get_oh() ?></span></td>
                                             </tr>
                                         </tbody>
                                     </table>
@@ -275,19 +187,19 @@ if (!isset($_COOKIE['email'])) { // If no cookie is present, redirect:
                                         <tbody>
                                             <tr>
                                             <td>Case Manager</td>
-                                            <td><span class="po" name="po-cm"><?php echo get_cost('po',4) ?></span></td>                                                
+                                            <td><span class="po" name="po-cm"><?php echo get_patient_cost('po',4) ?></span></td>
                                             </tr>    
                                             <tr>
                                             <td>Physician Assistant</td>
-                                            <td><span class="po" name="po-pa"><?php echo get_cost('po',10) ?></span></td>
+                                            <td><span class="po" name="po-pa"><?php echo get_patient_cost('po',10) ?></span></td>
                                             </tr>                                             
                                             <tr>
                                             <td>Registered Nurse</td>
-                                            <td><span class="po" name="po-rn"><?php echo get_cost('po',8) ?></span></td>                                                     
+                                            <td><span class="po" name="po-rn"><?php echo get_patient_cost('po',8) ?></span></td>
                                             </tr>
                                             <tr>
                                             <td>Respiratory Therapist</td>
-                                            <td><span class="po" name="po-rt"><?php echo get_cost('po',12) ?></span></td>                                                     
+                                            <td><span class="po" name="po-rt"><?php echo get_patient_cost('po',12) ?></span></td>
                                             </tr>
                                             <tr>
                                                 <td><strong>Total:</strong></td>
@@ -352,31 +264,31 @@ if (!isset($_COOKIE['email'])) { // If no cookie is present, redirect:
                                         <tbody>
                                             <tr>
                                             <td>Anesthesiologist</td>
-                                            <td><span class="to" name="to-an"><?php echo get_cost('to',1) ?></span></td>                                                     
+                                            <td><span class="to" name="to-an"><?php echo get_patient_cost('to',1) ?></span></td>
                                             </tr>
                                             <tr>
                                             <td>Cardiovascular Surgeon</td>
-                                            <td><span class="to" name="to-cs"><?php echo get_cost('to',3) ?></span></td>                                                     
+                                            <td><span class="to" name="to-cs"><?php echo get_patient_cost('to',3) ?></span></td>
                                             </tr>                                            
                                             <tr>
                                             <td>Case Manager</td>
-                                            <td><span class="to" name="to-cm"><?php echo get_cost('to',4) ?></span></td>                                                     
+                                            <td><span class="to" name="to-cm"><?php echo get_patient_cost('to',4) ?></span></td>
                                             </tr>    
                                             <tr>
                                             <td>Physician Assistant</td>
-                                            <td><span class="to" name="to-pa"><?php echo get_cost('to',10) ?></span></td>
+                                            <td><span class="to" name="to-pa"><?php echo get_patient_cost('to',10) ?></span></td>
                                             </tr>                                             
                                             <tr>
                                             <td>Registered Nurse</td>
-                                            <td><span class="to" name="to-rn"><?php echo get_cost('to',8) ?></span></td>                                                     
+                                            <td><span class="to" name="to-rn"><?php echo get_patient_cost('to',8) ?></span></td>
                                             </tr>
                                             <tr>
                                             <td>Respiratory Therapist</td>
-                                            <td><span class="to" name="to-rt"><?php echo get_cost('to',12) ?></span></td>                                                     
+                                            <td><span class="to" name="to-rt"><?php echo get_patient_cost('to',12) ?></span></td>
                                             </tr> 
                                             <tr>
                                             <td>Scrub Tech</td>
-                                            <td><span class="to" name="to-st"><?php echo get_cost('to',13) ?></span></td>
+                                            <td><span class="to" name="to-st"><?php echo get_patient_cost('to',13) ?></span></td>
                                             </tr>
                                             <tr>
                                                 <td><strong>Total:</strong></td>
@@ -491,7 +403,7 @@ if (!isset($_COOKIE['email'])) { // If no cookie is present, redirect:
             // get patient total cost
             var p_cost = (parseFloat($("#to-result").text().replace(/\,/g, '')) + parseFloat($("#to-result2").text()) + parseFloat($("#to-result3").text())).toFixed(0);
             // get avg cost
-            var h_cost = parseFloat("<?php echo get_value('cost', 'hospital'); ?>").toFixed(0);
+            var h_cost = parseFloat("<?php echo get_patient_value('cost', 'hospital'); ?>").toFixed(0);
             var data = [p_cost, h_cost, 6000];
             var dataLabel = ["Patient Cost","Average Cost (Hospital)","Average Reimbursement"];
             // y-axis
@@ -549,8 +461,8 @@ if (!isset($_COOKIE['email'])) { // If no cookie is present, redirect:
         
         function timeChartFunction() {
             // 1. Time Table
-            var p_time = parseFloat("<?php echo get_value('time', 'patient'); ?>").toFixed(1);
-            var h_time = parseFloat("<?php echo get_value('time', 'hospital'); ?>").toFixed(1);
+            var p_time = parseFloat("<?php echo get_patient_value('time', 'patient'); ?>").toFixed(1);
+            var h_time = parseFloat("<?php echo get_patient_value('time', 'hospital'); ?>").toFixed(1);
             var data = [p_time, h_time, 4.2];
             var dataLabel = ["Patient Stay","Average Stay (Hospital)","Average Stay (National)"];
             var barWidth = 30;
